backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# OpenAI and LangChain - Stable tested versions
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.8
openai==1.35.7
httpx==0.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Vector store and embeddings
faiss-cpu==1.7.4
tiktoken==0.8.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.0
markdown==3.5.2

# Data visualization
plotly==5.18.0
pandas==2.2.0

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==23.2.1
